requirements.txt
python=3.12
flask=3.0.3
flask-sqlalchemy= 3.1.1
flask-migrate=4.0.7
flask-wtf=1.2.1
psycopg2=2.9.9
python-dotenv=1.0.1
flask-login=0.6.3
email-validator=2.1.1